Queue Management | Concatenate Array in c++
We continue to develop our Queue management system that we made in the previous module.
You are asked to add new functionality: adding two queues together. The result should be a new queue, where the elements of the first queue come first, followed by the second queue's elements.

Given the Queue class, overload the + operator, so that the code in main works and successfully adds two queues.