Queue Management | Template Class
Class with string and integer || Template

It's time to update your Queue management system. The previous version supports only integer numbers and we need to support more types, such as strings, to store customer names in the queue.

Transform the given Queue class to a class template, which can work with different data types.